% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oz.R
\name{oz}
\alias{oz}
\title{The oz function}
\description{
The classic `oz` package `oz` function.
}
\details{
This function calls `oz::oz()` to draw a basic outline. See
`oz_data` to obtain the data in native form.

See \code{oz::\link[oz]{oz}} for details.
}
\keyword{internal}
