% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozmap.R
\name{ozmap}
\alias{ozmap}
\title{Australia map}
\usage{
ozmap(states = TRUE, add = FALSE, ...)
}
\arguments{
\item{states}{include state boundaries, `TRUE` by default}

\item{add}{add to existing plot, `FALSE` by default}

\item{...}{arguments passed to `oz::oz`}
}
\value{
nothing
}
\description{
Draw a map of Australia, with or without states.
}
\details{
This function calls `oz::oz()` to draw a basic outline. The
outline data is purely in longitude-latitude form, see
`ozmap_data()` to obtain the data itself.`
}
\examples{
ozmap(states = FALSE, col = "firebrick", lwd = 4)
ozmap(states = TRUE, add = TRUE)

## further arguments may be passed to oz::oz, like coast
ozmap(states = FALSE, col = "dodgerblue", lwd = 4)
ozmap(states = TRUE, add = TRUE, coast = FALSE)
}
\seealso{
ozmap_data
}
