% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozmap.R
\name{ozmap_data}
\alias{ozmap_data}
\alias{oz_data}
\title{Australia map data}
\usage{
ozmap_data(data = "states", quiet = FALSE, ...)
}
\arguments{
\item{data}{name of dat to return, see details}

\item{quiet}{set to `TRUE` to suppress messages}

\item{...}{passed to data wrapper functions, notably `oz::oz` from `oz_data`}
}
\value{
`sf` data frame
}
\description{
Australia map data
}
\examples{
ozmap_data("country")
}
