% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozmaps-package.R
\docType{data}
\name{ozmap_states}
\alias{ozmap_states}
\alias{ozmap_country}
\title{Australian map data}
\description{
Australian coastline and boundaries data, including states and territories
}
\details{
In-built data set of Australian coastline and provinces (states and territories) from Natural Earth.
These are selective versions of the `rnaturalearth` data sets.
`ozmap_states` from `rnaturalearth::ne_states(country = "australia")` (`adm1_code`)
`ozmap_country` from `rnaturalearth::ne_countries(country = "australia")` (`adm0_a3`)
}
