% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oz.R
\name{oz}
\alias{oz}
\alias{oz_data}
\title{The oz function}
\usage{
oz_data(data = "states", ...)
}
\arguments{
\item{data}{character string, "states" provides state level else country level}

\item{...}{passed to \code{\link[oz:oz]{oz::ozRegion()}}}
}
\value{
oz class list of coordinates
}
\description{
The classic \code{oz} package \code{\link[oz:oz]{oz::oz()}} function.
}
\details{
This function calls \code{\link[oz:oz]{oz::oz()}} to draw a basic outline. Use
\code{\link[=oz_data]{oz_data()}} to obtain the data in native form.

See \code{\link[oz:oz]{oz::ozRegion()}} for more details.  Here \code{data} is treated as
an identifier, but only "states" or any other value is accepted. If not
"states", then country level is returned. Further arguments to \code{\link[oz:oz]{oz::ozRegion()}}
can be passed in via dots.
}
