% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozmap.R
\name{ozmap}
\alias{ozmap}
\title{Australia map}
\usage{
ozmap(x = "ozmap_country", ..., add = FALSE)
}
\arguments{
\item{x}{name of data set to use, default is `ozmap_country`}

\item{...}{arguments passed to ...}

\item{add}{add to existing plot, `FALSE` by default}
}
\value{
nothing
}
\description{
Draw a map of Australia, with or without states.
}
\details{
outline data is purely in longitude-latitude form, see
`ozmap_data()` to obtain the data itself.`

ozmap_country, ozmap_states
}
\seealso{
ozmap_data
}
