% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozmap.R
\name{ozmap_data}
\alias{ozmap_data}
\title{Australia map data}
\usage{
ozmap_data(data = "states", quiet = FALSE, ...)
}
\arguments{
\item{data}{name of dat to return, see details}

\item{quiet}{set to \code{TRUE} to suppress messages}

\item{...}{passed to data wrapper functions, notably \code{oz::oz()}}
}
\value{
\code{sf} data frame with 'NAME' and 'geometry' columns
}
\description{
Return simple features data frames of various Australian map layers.
}
\details{
Available layers are
\itemize{
\item states
\item country
\item \link{abs_ced}
\item \link{abs_lga}
\item \link{abs_ste}
}
}
\examples{
ozmap_data("country")
}
