% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozmap.R
\name{ozmap_data}
\alias{ozmap_data}
\title{Australia map data}
\usage{
ozmap_data(data = "states", quiet = FALSE, ...)
}
\arguments{
\item{data}{name of layer to return, see details}

\item{quiet}{set to \code{TRUE} to suppress messages}

\item{...}{unused}
}
\value{
\code{sf} data frame with 'NAME' and 'geometry' columns
}
\description{
Return simple features data frames of various Australian map layers.
}
\details{
Available layers are
\itemize{
\item \href{ozmap_states}{states} state and territories (low resolution)
\item \href{ozmap_country}{country} entire country (low resolution)
\item \link{abs_ced} country level electoral divisions
\item \link{abs_lga} local government areas
\item \link{abs_ste} state and territories
}
}
\examples{
ozmap_data("country")

ozmap_data("abs_lga")
}
