% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozmaps-package.R
\docType{data}
\name{ozmap_states}
\alias{ozmap_states}
\alias{ozmap_country}
\title{Australian map data}
\description{
Australian coastline and boundaries data, including states and territories
}
\details{
In-built data set of Australian coastline and provinces (states and
territories) simplified from the Australian Bureau of Statistics layer
\link{abs_ste}.
}
