% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozmaps-package.R
\docType{package}
\name{ozmaps-package}
\alias{ozmaps}
\alias{ozmaps-package}
\title{ozmaps: Australia Maps}
\description{
Draw maps of Australian coastline and administrative regions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mdsumner/ozmaps}
  \item Report bugs at \url{https://github.com/mdsumner/ozmaps/issues}
}

}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

Other contributors:
\itemize{
  \item Dianne Cook [contributor]
}

}
\keyword{internal}
