% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozmaps-package.R
\docType{package}
\name{ozmaps-package}
\alias{ozmaps}
\alias{ozmaps-package}
\title{ozmaps: Australia Maps}
\description{
Maps of Australian coastline and administrative regions. Data 
 can be drawn or accessed directly as simple features objects. Includes
 simple functions for country or state maps of Australia and in-built data
 sets of administrative regions from the Australian Bureau of Statistics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mdsumner/ozmaps}
  \item Report bugs at \url{https://github.com/mdsumner/ozmaps/issues}
}

}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

Other contributors:
\itemize{
  \item Dianne Cook [contributor]
  \item Edzer Pebesma (wrote the plot code for sf internalized here) [contributor]
}

}
\keyword{internal}
